+++ example/app/lib/helpers.rb
@@ -1,0 +1,3 @@
+module Pakyow::Helpers
+  # define methods here that are available from routes, bindings, etc
+end