+++ pakyow-core/lib/pakyow/core.rb
@@ -1,7 +1,6 @@
 require 'find'
 require 'rack'
 require 'rack/file'
-require 'logger'
 require 'cgi'
 
 require 'pakyow/support'