+++ pakyow-core/lib/pakyow/core/config/logger.rb
@@ -1,3 +1,6 @@
+require 'pakyow/core/logger/formatters/dev_formatter'
+require 'pakyow/core/logger/formatters/logfmt_formatter'
+
 Pakyow::Config.register :logger do |config|
   # whether or not pakyow should write to a log
   config.opt :enabled, true
@@ -5,30 +8,20 @@ Pakyow::Config.register :logger do |config|
   # the default level to log at
   config.opt :level, :debug
 
-  # whether or not pakyow should log to stdout
-  config.opt :stdout, true
+  # the formatter responsible for formatting request logs
+  config.opt :formatter, Pakyow::Logger::DevFormatter
 
-  # where the log file should be placed
-  config.opt :path, -> { File.join(Pakyow::Config.app.root, 'log') }
-
-  # the name of the log file
-  config.opt :filename, 'pakyow.log'
-
-  # whether or not the log file should be synced
-  config.opt :sync
-
-  # whether or not the log file should be flushed automatically
-  config.opt :auto_flush
-
-  # whether or not the log file should be colorized
-  config.opt :colorize
-end.env :development do |opts|
-  opts.sync = true
-  opts.auto_flush = true
-  opts.colorize = true
+  # where we should log to
+  config.opt :destinations, -> do
+    if Pakyow::Config.logger.enabled
+      [$stdout]
+    else
+      ["/dev/null"]
+    end
+  end
+end.env :test do |opts|
+  opts.destinations = []
 end.env :production do |opts|
-  opts.sync = false
-  opts.auto_flush = false
-  opts.stdout = false
-  opts.colorize = false
+  opts.formatter = Pakyow::Logger::LogfmtFormatter
+  opts.level = :info
 end