+++ pakyow-core/spec/unit/app_spec.rb
@@ -1,0 +1,58 @@
+require "spec_helper"
+require "pakyow/core/app"
+
+describe Pakyow::App do
+  let :app do
+    Pakyow::App.new
+  end
+
+  it "is hookable" do
+    expect(Pakyow::App.ancestors).to include(Pakyow::Support::Hookable)
+  end
+
+  describe "known events" do
+    describe "init" do
+      it "is known" do
+        expect(Pakyow::App.is_known_event?(:init)).to be true
+      end
+    end
+
+    describe "configure" do
+      it "is known" do
+        expect(Pakyow::App.is_known_event?(:configure)).to be true
+      end
+    end
+
+    describe "load" do
+      it "is known" do
+        expect(Pakyow::App.is_known_event?(:load)).to be true
+      end
+    end
+
+    describe "reload" do
+      it "is known" do
+        expect(Pakyow::App.is_known_event?(:reload)).to be true
+      end
+    end
+
+    describe "fork" do
+      it "is known" do
+        expect(Pakyow::App.is_known_event?(:fork)).to be true
+      end
+    end
+  end
+
+  describe "#forking" do
+    it "calls before: :fork hooks" do
+      expect(app).to receive(:call_hooks).with(:before, :fork)
+      app.forking
+    end
+  end
+
+  describe "#forked" do
+    it "calls after: :fork hooks" do
+      expect(app).to receive(:call_hooks).with(:after, :fork)
+      app.forked
+    end
+  end
+end