+++ pakyow-realtime/lib/pakyow/realtime/exceptions.rb
@@ -1,6 +1,5 @@
 module Pakyow
   module Realtime
-    class HandshakeError < Error; end
     class MissingMessageHandler < Error; end
   end
 end