+++ pakyow-realtime/lib/pakyow/realtime/ext/app.rb
@@ -1,0 +1,8 @@
+module Pakyow
+  class App
+    # @api private
+    def socket
+      Realtime::Context.new(self)
+    end
+  end
+end