+++ pakyow-test/lib/pakyow/test.rb
@@ -26,11 +26,11 @@ module Pakyow
     def self.setup(path = './app/setup')
       require path
 
-      Pakyow::App.after :match do
+      Pakyow::CallContext.after :match do
         @presenter = Pakyow::TestHelp::ObservablePresenter.new(@presenter)
       end
 
-      Pakyow::App.before :process do
+      Pakyow::CallContext.before :process do
         Pakyow::TestHelp::Realtime::ObservableMutator.instance.reset
       end