+++ pakyow-ui/lib/pakyow/ui.rb
@@ -1,1 +1,6 @@
+require 'pakyow/support'
+require 'pakyow/core'
+require 'pakyow/presenter'
+require 'pakyow/realtime'
+
 require 'pakyow/ui/base'